rename ConfigurationController
